Working implementation of VSSGP

+++ ghost/regression/SNRpenalty.py
@@ -10,5 +10,5 @@ def SEard( loghyp, params={'log_snr': np.log(1000), 'log_ls': np.log(100), 'log_
     log_sn = loghyp[:,-1]
     log_ll = loghyp[:,:-2]
     penalty = ((log_sf - log_sn)/log_snr)**p
-    penalty += (((log_ll - log_std)/log_ls)**p).sum(1, acc_dtype=loghyp.dtype)
+    penalty += (((log_ll - log_std)/log_ls)**p).sum(1)
     return penalty
